requirements.txt
pyyaml=3.12
numpy==1.13.3
scipy==0.19.1
scikit-learn==0.19.1
nose==1.3.7
matplotlib==2.1.0
pyyaml==3.12
pytest==3.5.0